requirements.txt
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pandas>=2.0.0
databricks-sdk>=0.1.0

